Use "32-bit" and "64-bit" instead of "x86" and "x64"
This isn't the right abstraction. We're only interested in the size of
"words" (as defined by OCaml) that the OCaml system building the library
is using.

x86 is the name of an instruction set and x64 (or x86_64) is an
extension of that set.
OCaml also support other architectures (eg. ARM) that have 32-bit and
64-bit variants and can be built with a word size that doesn't
correspond to the host architecture.